+++ src/PerfCpuType.cpp
@@ -21,10 +21,22 @@
 // if USE_PAPI is defined, compile this file with openmp option
 
 #ifdef DISABLE_MPI
-#include "mpi_stubs.h"
+	#define MPI_COMM_WORLD 0
+	#define MPI_INT  1
+	#define MPI_CHAR 2
+	#define MPI_DOUBLE 3
+	#define MPI_UNSIGNED_LONG 4
+	typedef int MPI_Comm;
+	typedef int MPI_Datatype;
+	typedef int MPI_Op;
+	typedef int MPI_Group;
+	#define MPI_SUCCESS true
+	#define MPI_SUM (MPI_Op)(0x58000003)
+	#include "mpi_stubs.h"
 #else
-#include <mpi.h>
+	#include <mpi.h>
 #endif
+
 #include <cmath>
 #include "PerfWatch.h"
 
@@ -659,5 +671,16 @@ void PerfWatch::outputPapiCounterLegend (FILE* fp)
 }
 
 
-
 } /* namespace pm_lib */
+
+
+#ifdef DISABLE_MPI
+	#undef MPI_COMM_WORLD
+	#undef MPI_INT
+	#undef MPI_CHAR
+	#undef MPI_DOUBLE
+	#undef MPI_UNSIGNED_LONG
+	#undef MPI_SUCCESS
+	#undef MPI_SUM
+#endif
+