+++ ChangeLog.md
@@ -3,6 +3,11 @@
 ## REVISION HISTORY
 
 ---
+- 2017-05-12 Version 5.6.2
+  -  fix to suppress printing Root Section(*)
+  -  move macros such as MPI_SUM from include/mpi_stubs.h to *.cpp source files
+
+---
 - 2017-03-30 Version 5.6.1
   - change _PM_WITHOUT_MPI_ to DISABLE_MPI due to maintain the consistency with other libraries